Move PSAMacroEnumerator to macro_collector
It's useful for more than test_psa_constant_names.
